func: preview schedule, print SER
view-enlisted-sched.php is the one loading data to page table display. needs a view in phpmyadmin first (not a true cursor due to problems reading and writing data on the database). NEEDS integration with previous confirm enrollment step for variables in query

preview-regform still needs additional styling and integration with previous step. also not a cursor due to problems like in other file. uses select and a hidden element for a layout of printing preview.