Add files via upload
used by cursor-view-enlisted-sched.php to function. just add this in the database, it assumes that the section is a column in the enlisted subjects_table and is part of the compound key with enlistment_id and semester_id.

TODO: see if making a stored procedure query is possible without running into many errors if needed talaga na may cursor rito.